feat: add class Users
