+++ src/utils/initialState.js
@@ -1,7 +1,7 @@
 import { AVATAR_DEFAULT } from './url-utils';
 
 const initialState = {
-    users: [],
+  users: [],
 };
 
 export { initialState };
